requirements.txt
numpy
scipy
opencv
pymaxflow
PIL
lear-gist